REFACTOR :: type 재활용
